// types.ts
export enum ClientStatus {
  EN_ATTENTE = 'EN_ATTENTE',
  ACTIF = 'ACTIF',
  SUSPENDU = 'SUSPENDU',
  BLOQUE = 'BLOQUE',
  REACTIVATION_EN_ATTENTE = 'REACTIVATION_EN_ATTENTE',
  INACTIF_EXCLU = 'INACTIF_EXCLU',
  EN_VERIFICATION_DOUBLON = 'EN_VERIFICATION_DOUBLON',
  REJETE = 'REJETE',
}

export enum LoanStatus {
  EN_ATTENTE = 'EN_ATTENTE',
  APPROUVE = 'APPROUVE',
  DEBLOQUE = 'DEBLOQUE',
  REMBOURSE = 'REMBOURSE',
  REJETE = 'REJETE',
  DEFAUT = 'DEFAUT',
}

export enum InstallmentStatus {
  EN_ATTENTE = 'EN_ATTENTE',
  PAYEE = 'PAYEE',
  EN_RETARD = 'EN_RETARD',
  IMPAYEE = 'IMPAYEE',
  EN_ATTENTE_CONFIRMATION = 'EN_ATTENTE_CONFIRMATION',
}